Reorganise code to simplify package hierarchy
